+++ git_well/git_squash_streaks.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 # PYTHON_ARGCOMPLETE_OK
 """
 This git-squash-streaks command
@@ -91,7 +91,6 @@ class SquashStreakCLI(scfg.DataConfig):
 
     # TODO: scriptconfig needs to be extended to handle these argparse
     # use-cases
-
     # dry = scfg.Value(True, alias=['force'], mutex_group=1, short_alias=['f'], help='opposite of --dry', isflag=True)
     # nargs=0)
     # quiet = scfg.Value(None, alias=['quiet'], mutex_group=2, short_alias=['q'], help='suppress output', nargs=0)
@@ -211,8 +210,6 @@ def find_pseudo_chain(head, oldest_commit=None, preserve_tags=True):
     import networkx as nx
 
     graph = git_nx_graph(head, oldest_commit, preserve_tags=preserve_tags)
-    # for idx, node in enumerate(nx.dfs_preorder_nodes(graph)):
-    #     print('{}, {!r}'.format(idx, graph.nodes[node]['commit'].message))
 
     sinks = {node for node in graph.nodes if len(graph.succ[node]) == 0}
     sources = {node for node in graph.nodes if len(graph.pred[node]) == 0}
@@ -221,10 +218,6 @@ def find_pseudo_chain(head, oldest_commit=None, preserve_tags=True):
 
     source = ub.peek(sources)
     sink = ub.peek(sinks)
-
-    # all_paths = list(nx.all_simple_paths(graph, source, sink))
-    # all_paths = list(map(ub.oset, all_paths))
-    # common_path = ub.oset.intersection(*all_paths)
 
     ugraph = nx.to_undirected(graph)
 
@@ -308,7 +301,6 @@ def git_nx_graph(head, oldest_commit=None, preserve_tags=False):
     Ignore:
         from xdev.util_networkx import write_network_text
         write_network_text(graph)
-
     """
     repo = head.repo
 
@@ -471,7 +463,6 @@ def find_chain(head, authors=None, preserve_tags=True, oldest_commit=None):
                     git commit -am "initial commit"
                     git branch -D old_master
                 '''))
-
             break
 
         if stop_object is not None:
@@ -627,7 +618,7 @@ def checkout_temporary_branch(repo, suffix='-temp-script-branch'):
     return temp_branchname
 
 
-def commits_between(repo, start, stop):
+def commits_between(repo, start, stop, start_inclusive=True):
     """
     Args:
         start (Commit): toplogically first (i.e. chronologically older) commit
@@ -669,14 +660,21 @@ def commits_between(repo, start, stop):
         >>> repo = git.Repo(repo_dpath)
         >>> stop = repo.head.commit
         >>> start = stop.parents[0].parents[0].parents[0].parents[0]
-        >>> commits = commits_between(repo, start, stop)
+        >>> commits = commits_between(repo, start, stop, start_inclusive=True)
         >>> assert commits[0] == stop
         >>> assert commits[-1] == start
         >>> assert len(commits) == 5
+        >>> commits2 = commits_between(repo, start, stop, start_inclusive=False)
+        >>> assert commits2[0] == commits[0]
+        >>> assert commits2[-1] == commits[-2]
+        >>> assert len(commits2) == 4
     """
     import binascii
     import git
-    argstr = '{start}^..{stop}'.format(start=start, stop=stop)
+    if start_inclusive:
+        argstr = '{start}^..{stop}'.format(start=start, stop=stop)
+    else:
+        argstr = '{start}..{stop}'.format(start=start, stop=stop)
     hexshas = repo.git.rev_list(argstr).splitlines()
     binshas = [binascii.unhexlify(h) for h in hexshas]
     commits = [git.Commit(repo, b) for b in binshas]
@@ -687,7 +685,7 @@ class RollbackError(Exception):
     pass
 
 
-def _squash_between(repo, start, stop, dry=False, verbose=True):
+def _squash_between(repo, start, stop, dry=False, verbose=True, start_inclusive=True):
     """
     inplace squash between, use external function that sets up temp branches to
     use this directly from the commandline.
@@ -714,7 +712,9 @@ def _squash_between(repo, start, stop, dry=False, verbose=True):
         ts_stop_short = ts_stop
 
     # Construct a new message
-    commits = commits_between(repo, start, stop)
+    commits = commits_between(repo, start, stop, start_inclusive=start_inclusive)
+    print(len(commits))
+    print(f'commits={commits}')
     messages = [commit.message for commit in commits]
     # messages = [commit.message for commit in streak._streak]
     unique_messages = ub.unique(messages)
@@ -726,6 +726,7 @@ def _squash_between(repo, start, stop, dry=False, verbose=True):
         new_msg = '{} - Squashed {} commits from <{}> to <{}>\n'.format(
             summary, len(commits), ts_start, ts_stop_short)
     else:
+        # TODO: need more options for messages
         new_msg = '{} - Squashed {} commits'.format(summary.strip(), len(commits))
 
     if verbose:
@@ -743,8 +744,11 @@ def _squash_between(repo, start, stop, dry=False, verbose=True):
         # ------------------
         # Go back in time to the sequence stopping point
         repo.git.reset(stop.hexsha, hard=True)
-        # Undo commits from start to stop by softly reseting to just before the start
-        before_start = start.parents[0]
+
+        # Undo commits from the first included commit after start to stop by
+        # softly reseting to just before the that included start commit.
+        start_included = commits[-1]
+        before_start = start_included.parents[0]
         if verbose:
             print(' * reseting to before <start>: {}'.format(before_start.hexsha))
         repo.git.reset(before_start.hexsha, soft=True)
@@ -768,19 +772,12 @@ def _squash_between(repo, start, stop, dry=False, verbose=True):
 
                 if EXPERIMENTAL_REBASE:
                     # above = streak.child.hexsha + '..' + old_head
-                    # repo.git.cherry_pick(above, allow_empty=True, mainline=1)
-
                     # git rebase --onto master topicA topicB
                     # git rebase --onto <current> <stop> <old_head>
                     repo.git.rebase(stop, old_head, preserve_merges=True, onto='HEAD')
-
-                    # git rebase --preserve-merges --onto dev/0.9.2-squash-temp 52f34a11b837d27f9979d002391c0d8d6bee4957 9e2c03c0df13d9ce3c74aa0ff619abe21afe51b1
-
-                    # repo.git.rebase(above, allow_empty=True, mainline=1)
                 else:
                     # Fixme, do this with rebase to preserve merges?
                     repo.git.cherry_pick(above, allow_empty=True)
-                # sys.exit(1)
             except git.GitCommandError:
                 print('ERROR: need to roll back')
                 raise
@@ -796,7 +793,6 @@ def do_tags(verbose=True, inplace=False, dry=True, auto_rollback=False):
             print('squashing streaks (DRY RUN)')
         else:
             print('squashing streaks')
-        # print('authors = {!r}'.format(authors))
 
     # If you are in a repo subdirectory, find the repo root
     cwd = os.getcwd()
@@ -812,7 +808,6 @@ def do_tags(verbose=True, inplace=False, dry=True, auto_rollback=False):
     repo = git.Repo(repodir)
     orig_branch_name = repo.active_branch.name
 
-    # head = repo.commit('HEAD')
     info = ub.cmd('git tag -l --sort=v:refname', verbose=3)
 
     info2 = ub.cmd('git show-ref --tags', verbose=3)
@@ -839,17 +834,14 @@ def do_tags(verbose=True, inplace=False, dry=True, auto_rollback=False):
         b = repo.commit(hash_b)
         if repo.is_ancestor(ancestor_rev=a, rev=b):
             a, b = b, a
-        # assert repo.is_ancestor(ancestor_rev=b, rev=a)
         streak = Streak(a, _streak=[a, b])
 
         if len(streak.start.parents) != 1:
             print('WARNING: cannot include streak = {!r}'.format(streak))
             continue
-        # assert start.authored_datetime < stop.authored_datetime
         if not repo.is_ancestor(ancestor_rev=streak.start, rev=streak.stop):
             print('WARNING: cannot include streak = {!r}'.format(streak))
             continue
-            # raise AssertionError('cant handle')
         streaks.append(streak)
 
     if verbose:
@@ -991,17 +983,12 @@ def squash_streaks(authors, timedelta='sameday', pattern=None,
                                oldest_commit=oldest_commit)
 
         if verbose:
-            # ISO_8601 = '%Y-%m-%d %H:%M:%S %z'  # NOQA
-            # ts_start = start.authored_datetime.
-            # print(ub.urepr([(c.message.strip(), c.author.name, c.authored_datetime.strftime(ISO_8601)) for c in chain]))
-            # print(ub.urepr(chain, nl=1))
             print('Found chain of length {!r}'.format(len(chain)))
 
         streaks = find_streaks(chain, authors=authors, timedelta=timedelta,
                                pattern=pattern)
     if verbose:
         print('Found %r streaks' % (len(streaks)))
-        # sys.exit(0)
 
     # Switch to a temp branch before we start working
     if not dry:
@@ -1014,8 +1001,6 @@ def squash_streaks(authors, timedelta='sameday', pattern=None,
             if verbose:
                 print('Squashing streak = %r' % (str(streak),))
             # Start is the commit further back in time
-            # print('streak = {!r}'.format(streak))
-            # print('streak.start = {!r}, {}'.format(streak.start, streak.start.message))
             _squash_between(repo, streak.start, streak.stop, dry=dry,
                             verbose=verbose)
     except Exception:
@@ -1024,7 +1009,6 @@ def squash_streaks(authors, timedelta='sameday', pattern=None,
         if not dry and auto_rollback:
             print('ROLLING BACK')
             repo.git.checkout(orig_branch_name)
-            # repo.git.branch(D=temp_branchname)
         print('You can debug the difference with:')
         print('    gitk {} {}'.format(orig_branch_name, temp_branchname))
         return
@@ -1051,95 +1035,10 @@ def squash_streaks(authors, timedelta='sameday', pattern=None,
             print('Or, to automatically accept changes run with --inplace')
 
 
-# def _autoparse_desc(func):
-#     try:
-#         # TODO: can we autogenerate the entire argument parser from the
-#         # docstring? or at least sectinons of it?
-#         # Parse docstrings for help strings
-#         from xdoctest.docstr import docscrape_google as scrape
-#         docstr = func.__doc__
-#         help_dict = {}
-#         for argdict in scrape.parse_google_args(docstr):
-#             help_dict[argdict['name']] = argdict['desc']
-#         description = scrape.split_google_docblocks(docstr)[0][1][0].strip()
-#         description = description.replace('\n', ' ')
-#     except ImportError:
-#         from collections import defaultdict
-#         help_dict = defaultdict(lambda: '')
-#         description = ''
-#     return description, help_dict
-
-
-# commandline entry point
 def git_squash_streaks(cmdline=1, **kwargs):
     """
     git-squash-streaks
-
-    Usage:
-        See argparse
     """
-    # import argparse
-    # try:
-    #     import argcomplete
-    # except Exception:
-    #     argcomplete = None
-    # description, help_dict = _autoparse_desc(squash_streaks)
-
-    # parser = argparse.ArgumentParser(description=description)
-    # parser.add_argument(*('--timedelta',), type=str,
-    #                     help=help_dict['timedelta'])
-
-    # parser.add_argument(*('--custom_streak',), nargs=2,
-    #                     help='hack to specify one custom streak: older newer')
-
-    # parser.add_argument(*('--pattern',), type=str,
-    #                     help=help_dict['pattern'])
-
-    # parser.add_argument(*('--tags',), action='store_true', help='experimental')
-
-    # parser.add_argument(*('--no-preserve-tags',), dest='preserve_tags',
-    #                     action='store_false', help=help_dict['preserve_tags'])
-
-    # parser.add_argument(*('--oldest-commit',), dest='oldest_commit',
-    #                     help=help_dict['oldest_commit'])
-
-    # parser.add_argument(*('--inplace',), action='store_true',
-    #                     help=help_dict['inplace'])
-
-    # parser.add_argument(*('--auto-rollback',), action='store_true',
-    #                     dest='auto_rollback', help=help_dict['auto_rollback'])
-
-    # parser.add_argument('--authors', type=str,
-    #                     help=(help_dict['authors'] +
-    #                           ' Only squash commits from these authors. '
-    #                           ' Set to <config> to use your git config'))
-
-    # group = parser.add_mutually_exclusive_group()
-    # group.add_argument(*('-n', '--dry'), dest='dry', action='store_true',
-    #                     help=help_dict['dry'])
-    # group.add_argument(*('-f', '--force'), dest='dry', action='store_false',
-    #                     help='opposite of --dry')
-
-    # group = parser.add_mutually_exclusive_group()
-    # group.add_argument(*('-v', '--verbose'), dest='verbose', action='store_const',
-    #                    const=1, help='verbosity flag flag')
-    # group.add_argument(*('-q', '--quiet'), dest='verbose', action='store_const',
-    #                    const=0, help='suppress output')
-
-    # parser.add_argument(*('--inplace',), action='store_true',
-    #                     help=help_dict['inplace'])
-    # parser.set_defaults(
-    #     tags=False,
-    #     inplace=False,
-    #     preserve_tags=True,
-    #     auto_rollback=False,
-    #     authors=None,
-    #     pattern=None,
-    #     timedelta='sameday',
-    #     dry=True,
-    #     verbose=True,
-    # )
-
     args = SquashStreakCLI.cli(cmdline=cmdline, data=kwargs, strict=True)
     ns = dict(args).copy()
     if ns.pop('tags'):
@@ -1205,7 +1104,5 @@ if __name__ == '__main__':
         git-squash-streaks --oldest-commit=master --timedelta=None
 
         git-squash-streaks --timedelta=None
-
-
     """
     main()